be/requirements.txt
fastapi
uvicorn
chromadb
google-generativeai
pymupdf
openpyxl
beautifulsoup4
chardet
python-docx
pyyaml
msal
requests
beautifulsoup4
httpx
pandasai[excel]
pandas
numpy
pdfplumber
langchain-text-splitters
langchain-core